fix: support fast checking dts files
